pyproject.toml
[project]
name = "arash-bot"
version = "1.0.0"
description = "Arash External API Service - Multi-platform AI chatbot with team-based access control"
authors = [
    {name = "Arash Team", email = "team@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Web Framework
    "fastapi==0.118.0",
    "uvicorn==0.37.0",
    "starlette==0.48.0",

    # HTTP Client
    "httpx==0.28.1",
    "httpcore==1.0.9",
    "h11==0.16.0",

    # Configuration & Validation
    "pydantic==2.11.9",
    "pydantic-settings==2.11.0",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.1",

    # Telegram Bot
    "python-telegram-bot==22.5",

    # Database
    "sqlalchemy==2.0.36",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "alembic==1.14.0",
    "greenlet==3.1.1",

    # Utilities
    "click==8.3.0",
    "colorama==0.4.6",
    "certifi==2025.8.3",
    "idna==3.10",
    "sniffio==1.3.1",
    "anyio==4.11.0",
    "annotated-types==0.7.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tabulate==0.9.0",
    "jdatetime>=5.2.0,<6.0.0",
    "pytz>=2024.2,<2025.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.8.0,<1.0.0",
    "mypy>=1.13.0,<2.0.0",
]

[project.scripts]
migrate = "scripts.migrate:main"
manage-keys = "scripts.manage_api_keys:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "telegram_bot*", "scripts*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
